Refactored lots of code, moved to blob_types from manual net_load()/net_store() functions

+++ main/log_utils.h
@@ -59,6 +59,6 @@ constexpr bool is_log_feature_present(LogFeatures::Enum components) {
 
 #define write_log(address, feature, string, ...) do {            \
 if constexpr (is_log_feature_present(feature)) {                 \
-printf("[%u]: " string "\n", address __VA_OPT__(,) __VA_ARGS__); \
+printf("[%u]: " string "\n", (uint) address __VA_OPT__(,) __VA_ARGS__); \
 fflush(stdout);                                                  \
 } } while(0)
